+++ src/App.tsx
@@ -1,11 +1,9 @@
 import { Route, Routes } from 'react-router-dom';
-import AddTime from './pages/AddTime';
-import Current from './pages/Current';
+import AddTime from './pages/addTime/AddTime';
 
 function App() {
   return (
     <Routes>
-      <Route path="/current" element={<Current />} />
       <Route path="/addTime" element={<AddTime />} />
     </Routes>
   );
