+++ src/components/addTable/AddTable.tsx
@@ -1,10 +1,20 @@
-import styled from '@emotion/styled';
-import theme from '../../styles/theme';
+import {
+  Blank,
+  Bottom,
+  Date,
+  DateWrapper,
+  Select,
+  SelectWrapper,
+  Time,
+  TimeWrapper,
+  Top,
+  Wrapper,
+} from './addTable.styles';
 
 interface Props {
   selectedMethod: string;
   tablePage: number;
-  validDateChunks: any;
+  validDateChunks: Array<{ date: string; isValidDate: boolean }[]>;
 }
 
 const AddTable = ({ selectedMethod, tablePage, validDateChunks }: Props) => {
@@ -42,11 +52,6 @@ const AddTable = ({ selectedMethod, tablePage, validDateChunks }: Props) => {
     '23:30',
   ];
 
-  const handleClick = (e: any) => {
-    console.log(e.currentTarget.getAttribute('value'));
-    console.log('class: ', e.currentTarget);
-  };
-
   return (
     <Wrapper>
       <Top>
@@ -75,7 +80,6 @@ const AddTable = ({ selectedMethod, tablePage, validDateChunks }: Props) => {
                 value={`${date} ${time}`}
                 isSelected={false}
                 selectedMethod={selectedMethod}
-                onClick={handleClick}
                 isValidDate={isValidDate}
               />
             ))}
@@ -85,113 +89,5 @@ const AddTable = ({ selectedMethod, tablePage, validDateChunks }: Props) => {
     </Wrapper>
   );
 };
-
-const Wrapper = styled.div`
-  width: 231px;
-  border-radius: 5.5px;
-  border: 1px solid ${theme.colors.gray03};
-
-  color: ${theme.colors.gray06};
-  ${theme.typography.medium02};
-`;
-
-const Top = styled.div`
-  display: flex;
-  width: 100%;
-  height: 37px;
-  border-bottom: 1px solid ${theme.colors.gray03};
-`;
-
-const Bottom = styled.div`
-  display: flex;
-  width: 100%;
-`;
-
-const Blank = styled.div`
-  width: 17px;
-  height: 36px;
-
-  border-right: 1px solid ${theme.colors.gray03};
-`;
-
-const DateWrapper = styled.div`
-  display: flex;
-`;
-
-const Date = styled.div`
-  width: 70px;
-  height: 36px;
-
-  display: flex;
-  align-items: center;
-  justify-content: center;
-
-  color: ${theme.colors.gray06};
-  ${theme.typography.medium04};
-
-  & + & {
-    width: 71px;
-    border-left: 1px solid ${theme.colors.gray03};
-  }
-`;
-
-const TimeWrapper = styled.div`
-  width: 17px;
-  border-right: 1px solid ${theme.colors.gray03};
-`;
-
-const Time = styled.div`
-  width: 16px;
-  height: 36px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-
-  color: ${theme.colors.gray04};
-  ${theme.typography.regular03};
-
-  & + & {
-    height: 37px;
-    border-top: 1px solid ${theme.colors.gray03};
-  }
-`;
-
-const SelectWrapper = styled.div`
-  width: 70px;
-
-  & + & {
-    width: 71px;
-    border-left: 1px solid ${theme.colors.gray03};
-  }
-`;
-
-const Select = styled.div<{
-  isSelected: boolean;
-  value: string;
-  selectedMethod: string;
-  isValidDate: boolean;
-}>`
-  height: 17px;
-  box-sizing: content-box;
-
-  background: ${({ isValidDate }) => !isValidDate && `${theme.colors.gray02}`};
-
-  background: ${({ selectedMethod, isSelected }) =>
-    isSelected &&
-    (selectedMethod ? `${theme.colors.purple06}` : `${theme.colors.orange02}`)};
-
-  &:nth-of-type(odd) {
-    border-bottom: 1px dashed ${theme.colors.gray03};
-  }
-
-  &:nth-of-type(even) {
-    height: 18px;
-    border-bottom: 1px solid ${theme.colors.gray03};
-  }
-
-  &:last-of-type {
-    border-bottom: none;
-  }
-`;
 
 export default AddTable;
