+++ src/utils/getChunks.ts
@@ -1,5 +1,10 @@
-export const getChunks = (validDates: any) => {
-  let dateChunks: any = [];
+interface ValidDate {
+  date: string;
+  isValidDate: boolean;
+}
+
+export const getChunks = (validDates: ValidDate[]) => {
+  let dateChunks: Array<ValidDate[]> = [];
 
   for (let i = 0; i < validDates.length; i += 3) {
     dateChunks = [...dateChunks, validDates.slice(i, i + 3)];
