+++ src/utils/getValidDates.ts
@@ -1,14 +1,18 @@
 import room from '../assets/data/room.json';
+interface ValidDate {
+  date: string;
+  isValidDate: boolean;
+}
 
 export const getValidDates = (dates: string[]) => {
-  let validDates: any = [];
+  let validDates: Array<ValidDate[]> = [];
 
   validDates = [
     ...validDates,
     dates.map((date) =>
-      room.dates.indexOf(date) === -1
-        ? { date: date, isValidDate: false }
-        : { date: date, isValidDate: true }
+      room.dates.includes(date)
+        ? { date: date, isValidDate: true }
+        : { date: date, isValidDate: false }
     ),
   ];
